refactor, add course page, delete some legacy code
