+++ constants/sampleCourses.tsx
@@ -1,0 +1,12 @@
+import { Course } from "@/definitions/course";
+
+export const fakeCourses: Course[] = [
+  { id: "1", name: "Pebble Beach", location: "Monteray, CA" },
+  { id: "2", name: "Augusta National", location: "Augusta, GA" },
+  { id: "3", name: "Baltusrol Golf Club", location: "Springfield, NJ" },
+  { id: "4", name: "Bethpage Black", location: "Bethpage, NY" },
+  { id: "5", name: "Pine Valley", location: "Pine Valley, NJ" },
+  { id: "6", name: "Shinnecock Hills", location: "Southampton, NY" },
+  { id: "7", name: "Oakmont", location: "Oakmont, PA" },
+  { id: "8", name: "Merion (East)", location: "Admore, PA" },
+];
