+++ definitions/course.tsx
@@ -1,0 +1,5 @@
+export type Course = {
+  id: string;
+  name: string;
+  location: string;
+};
